Interprete command substitution correctly in Bash
When running install command in MacOS Catalina (Bash version 3.2.57),
the install command fails with the error:

    .asdf/plugins/jq/bin/install: line 24: syntax error near unexpected token `;;'

This is because in the platform function the command substitution is
misinterpreted due to the closing parentheses of the case patterns.

Prefixing the case patterns with opening parenthenses ensures that the
command substitution is interpreted correctly.